Version 0.6.2.(2) Change Log:  - Product class is now abstract.  - Big fixes related to misspelled method names/signatures in override routines.  - PROJECT rules change - As of now @Override annotation is absolutely mandatory!  - Adding new functionalities for invoking discounts in certain situations.  - Adding necessary overloaded constructors to fulfil the needs of applyRating()'s calls to constructors not present by now.
Since adding new types of products can lead to potential problems when trying to use "applyRating()" that might remain not overridden in the future subtypes, the applyRating() method in the Product class is now set as abstract, as a result Product becomes abstract all subclasses that extend Product shall govern their very own version of applyRating();

Food class now overrides getDiscount() based on expiration date. If expiration date matches is today a discount is returned to the caller, if the current calendar day does not much the expiration date of the Food product BigDecimal.ZERO is returner;

Drink class now overrides getDiscount() by checking the current local time for particular time frame called Happy Hour. Happy Hour is hardcoded in the body for the getDiscount(). If the  local time when getDiscount() is called is withing the happy hour time frame, then the default getDiscount method of Product is invoked, else BigDecimal.ZERO is returned to the caller;